Separated building card display from building card script
