+++ Assets/Prefabs/UI/Building Cards.prefab
@@ -109,8 +109,6 @@ MonoBehaviour:
   layerMask:
     serializedVersion: 2
     m_Bits: 83886081
-  tweenTime: 0.5
-  tweenEase: 0
   particle: {fileID: 2873902488680928939, guid: 2f66426f4a963044f985cd4b94727a4c,
     type: 3}
   container: {fileID: 0}
@@ -173,6 +171,16 @@ PrefabInstance:
         type: 3}
       propertyPath: m_Name
       value: Building Card 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5223952580950037818, guid: e0cfdd53e09f042388fafc21800681d7,
+        type: 3}
+      propertyPath: m_Group
+      value: 
+      objectReference: {fileID: 7747285992698156223}
+    - target: {fileID: 5223952580950037818, guid: e0cfdd53e09f042388fafc21800681d7,
+        type: 3}
+      propertyPath: onValueChanged.m_PersistentCalls.m_Calls.Array.data[0].m_Mode
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 6579424632611495478, guid: e0cfdd53e09f042388fafc21800681d7,
         type: 3}
@@ -426,6 +434,11 @@ PrefabInstance:
       propertyPath: m_Name
       value: Building Card 3
       objectReference: {fileID: 0}
+    - target: {fileID: 5223952580950037818, guid: e0cfdd53e09f042388fafc21800681d7,
+        type: 3}
+      propertyPath: m_Group
+      value: 
+      objectReference: {fileID: 7747285992698156223}
     - target: {fileID: 6579424632611495478, guid: e0cfdd53e09f042388fafc21800681d7,
         type: 3}
       propertyPath: m_LocalRotation.w
@@ -673,6 +686,11 @@ PrefabInstance:
       propertyPath: m_Name
       value: Building Card 2
       objectReference: {fileID: 0}
+    - target: {fileID: 5223952580950037818, guid: e0cfdd53e09f042388fafc21800681d7,
+        type: 3}
+      propertyPath: m_Group
+      value: 
+      objectReference: {fileID: 7747285992698156223}
     - target: {fileID: 5288245882272591112, guid: e0cfdd53e09f042388fafc21800681d7,
         type: 3}
       propertyPath: m_AnchoredPosition.y
