+++ Assets/Prefabs/UI/BuildingCardDisplay.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b7d2f3f91f10c405286e87e6c4a9f595
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
