+++ Assets/Scripts/UI/BuildingCardDisplay.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e39f12ea1d5e24e29aa49a5454ff6370
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
